pyproject.toml
[project]
name = "crane-x7-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynamixel-sdk>=3.7.31",
    "numpy>=2.3.2",
    "pyyaml>=6.0.2",
]

[[tool.mypy.overrides]]
module = [
  "dynamixel_sdk.*",
  "numpy.*",
  "yaml.*",
]
ignore_missing_imports = true
